Changed grid coords to int,int
Grid coords will never be big enough to justify quint64

+++ src/selection.cc
@@ -1,23 +1,23 @@
 #include "selection.h"
 #include <QPoint>
 
-void Selection::start(quint64 x, quint64 y, quint64 tell)
+void Selection::start(int x, int y, quint64 tell)
 {
 	start_offset = xy_to_offset(x, y, tell);
 	end_offset = start_offset;
 }
 
-void Selection::end(quint64 x, quint64 y, quint64 tell)
+void Selection::end(int x, int y, quint64 tell)
 {
 	end_offset = xy_to_offset(x, y, tell);
 }
 
-quint64 Selection::xy_to_offset(quint64 x, quint64 y, quint64 tell)
+quint64 Selection::xy_to_offset(int x, int y, quint64 tell)
 {
 	return tell + (columns * y) + x;
 }
 
-bool Selection::in_range(quint64 x, quint64 y, quint64 tell)
+bool Selection::in_range(int x, int y, quint64 tell)
 {
 	quint64 test = xy_to_offset(x, y, tell);
 	quint64 s = start_offset;
